intersect_world

+++ source/intersection/s_intersection/si_size.c
@@ -6,13 +6,13 @@
 /*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 13:59:36 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/02/20 21:20:12 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/02/22 13:25:02 by tda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <header.h>
 
-int	si_size(t_intersection *lst)
+size_t	si_size(t_intersection *lst)
 {
 	t_intersection	*nextl;
 	size_t			i;
